+++ crates/icy_board_engine/src/vm/tests/mod.rs
@@ -24,6 +24,7 @@ mod fonts;
 mod forward_calls;
 mod graphics;
 mod hyperlinks;
+mod long;
 mod margins;
 mod masks;
 mod member_calls;